fix(design): run the REAL audit in the self-check, and move the numbers out of the components
Three review threads, each verified before changing anything.

**The self-check was not checking the checker.** `self-check.mjs` never called
`audit()` — it re-implemented a simplified version of the measurement inline, so
a regression INSIDE the audit passed it silently. That is exactly how this audit
has broken twice before (the on-screen filter that excluded the elements
`CLIPPED_BY_CONTAINER` exists to find; the rgb colour parsing that turned 189
real contrast failures into noise).

It could not import the audit because `layout-audit.mjs` launches Chromium at
import time. So the audit now lives alone in `tests/design/audit-fn.mjs`,
side-effect free, and BOTH callers run that one copy. The self-check asserts
finding TYPES from the real audit: `CLIPPED_BY_CONTAINER` appears when the
controls are clipped and not otherwise.

Honest gap, stated in the code and in the new README: the audit has NO
ragged-height finding, so that assertion cannot come from it. It is kept as a
separate measurement and labelled "page:" rather than dressed up as a check of
the checker.

**Context7 first for App Router work.** Added to `frontend/AGENTS.md`, but
OUTSIDE the `BEGIN/END:nextjs-agent-rules` block — `next dev` rewrites
everything inside it (`generate-agent-files.js`), so the suggested in-block edit
would be erased on the next run.

**The long comments.** Agreed on the principle: viewport measurements do not
belong in a component. The WHY stays in the code in one or two lines; the
numbers (189/231/191px, 79px/116px, 390px, the 9 broken tests) move to
`frontend/tests/design/README.md`, which is where the tool that produced them
lives. Nothing is lost, and a number in a README rots more loudly than one in a
JSX comment.

Verified: `node --check` passes on all three .mjs files, and the component diff
is comment-only — no className, no JSX, no logic changed (proof: every changed
line in the two .tsx files is inside a comment).
