+++ frontend/src/components/jobs/JobCard.tsx
@@ -159,11 +159,8 @@ export function JobCard({ job, onAction }: JobCardProps) {
           router.push(`/jobs/${job.id}`);
         }
       }}
-      // h-full: the cards sit in a 2/3-column grid, but each one is wrapped in a
-      // plain block div, so the CARD was never the grid item and never got the
-      // row's stretched height. Measured at 1440px, one row of three came out
-      // 189 / 231 / 191px tall — ragged bottoms and three different gaps before
-      // the next row. Filling the wrapper makes every card in a row equal.
+      // h-full: the card is wrapped in a plain div, so it is not the grid item
+      // and never inherits the row's stretched height. Numbers: tests/design/README.md.
       className="glass-card flex h-full cursor-pointer flex-col gap-3 rounded-xl p-4"
       role="link"
       tabIndex={0}
@@ -434,17 +431,10 @@ export function JobCard({ job, onAction }: JobCardProps) {
         ))}
       </div>
 
-      {/* ---- Actions ----
-          flex-wrap: this row holds Apply, Tailor my CV, Like, Skip, Details and
-          the source tag. With no wrapping it simply ran past the card, and the
-          card is overflow:hidden, so the overspill was CLIPPED AWAY rather than
-          scrolled — measured at 1440px, the Details button sat 79px and the
-          source label 116px beyond the padding box, i.e. both were invisible
-          and unclickable. Wrapping shows them instead of hiding them.
-
-          mt-auto: with h-full above, this pins the actions to the bottom of the
-          card so the button rows line up across a row of cards instead of
-          floating at whatever height the title happened to end. */}
+      {/* Actions. flex-wrap: the card is overflow:hidden, so without wrapping the
+          last controls are clipped away rather than scrolled — invisible AND
+          unclickable. mt-auto pins this row to the bottom so buttons line up
+          across a row of cards. Numbers: tests/design/README.md. */}
       <div className="mt-auto flex flex-col gap-2 border-t border-border/50 pt-3">
         {/* Row 1 — the two things you DO with a job. Equal width so the pair
             reads as one control and lands on the same baseline on every card. */}
